Refactor enhanced logging setup and telemetry collection features
