+++ core/setup_logging.py
@@ -1,0 +1,22 @@
+import logging
+import sys
+from visual.colored_formatter import ColoredFormatter
+
+def setup_logging():
+    """Setup enhanced logging with colors and formatting"""
+    logger = logging.getLogger()
+    logger.setLevel(logging.INFO)
+
+    # Clear existing handlers
+    for handler in logger.handlers[:]:
+        logger.removeHandler(handler)
+
+    # Console handler with colors
+    console_handler = logging.StreamHandler(sys.stdout)
+    console_handler.setFormatter(ColoredFormatter(
+        "[🔥 HexStrike AI] %(asctime)s [%(levelname)s] %(message)s",
+        datefmt="%Y-%m-%d %H:%M:%S"
+    ))
+    logger.addHandler(console_handler)
+
+    return logger
